+++ src/main/java/scanner/brute/RTSPMode.java
@@ -1,5 +1,14 @@
 package scanner.brute;
 
+/**
+ * Enum of possible RTSP request state.<br/>
+ * <ul>
+ * <li>{@link RTSPMode#ORTHODOX} - attack will be carry out by root path.</li>
+ * <li>{@link RTSPMode#SPECIAL} - attack will be carry out by special path</li>
+ * </ul>
+ *
+ * @author inkarnadin
+ */
 public enum RTSPMode {
 
     ORTHODOX,